刪除跳回本頁

+++ proj/ab-list.php
@@ -92,7 +92,11 @@ if ($totalRows > 0) {
             <?php foreach ($rows as $r) : ?>
                 <tr>
                     <td>
-                        <a href="ab-delete.php?sid=<?= $r['sid'] ?>">
+                        <?php /* <a href="ab-delete.php?sid=<?= $r['sid'] ?>" onclick="return confirm('確定要刪除編號為<?= $r['sid'] ?>的資料嗎?)">
+                            <i class="fa-solid fa-trash-can"></i>
+                        </a>
+                        */ ?>
+                        <a href="javascript: delete_it(<?= $r['sid'] ?>)">
                             <i class="fa-solid fa-trash-can"></i>
                         </a>
                     </td>
@@ -119,4 +123,14 @@ if ($totalRows > 0) {
 
 
 <?php include __DIR__ . '/parts/scripts.php' ?>
+<script>
+    function delete_it(sid) {
+        if (confirm(`確定要刪除編號為${sid}的資料嗎??`)) {
+            location.href = `ab-delete.php?sid=${sid}`
+        }
+    }
+</script>
+
+
+
 <?php include __DIR__ . '/parts/html-foot.php' ?>
